Fixes for typographic clock string.
Need plural forms for some languages based on the number of hours.

Need template where "hours" and "minute" is filled into the sentence.
This enables a conjunction to be added between "hours" and "minutes".
And enables, for languages like Basque, the verb "It's" to moved from
the beginning. "It's" can be moved either between "hours" and "minutes"
or after "minutes".

Bug: 123638383
Bug: 123638464
Bug: 123643485
Test: mp sysuig and manually tested
Change-Id: I69865d004ed346a21a13e86d96a4bdad379f08df