Restrict access to mRequestedNumColumns.
There is little value in exposing this as public API as general users
of the GridView have this value as they had to set it via setNumColumns.
Also classes extending GridView can keep track of this value by
overriding setNumColumns.

At the same time I don't see much motivation in needing the value as it
is only used in private method determineColumns.

Also the amount of usage of this field is really low.

Bug: b/123769395
Test: N/A
Change-Id: I0996cd1df2d1c3394b17333762a19a057005137a