Log display IDs for both IME and IME-client in StartInputHistory
This is a follow up CL to a recent CL [1], which added a special
fallback mechanism for IMEs to be shown on the default display when
the target display doesn't have FLAG_CAN_SHOW_SYSTEM_DECORATIONS.

It turns out that this scenario will soon become more user-visible
once Bubble UI is enabled, because Bubble UI is built on top of
ActivityView, where the target Activity including any text input field
will be running on a virtual display, not on the default display,
while IMEs are expected to continue running on the default display.

To make it easier to diagnose this new scenario, this CL updates our
retrospective debugging mechanism called StartInputHistory [2].  With
this CL, InputMethodManagerService starts keeping track of both IME
display ID and IME client display ID for each StartInput so that we
can see them later into the StartInputHistory section of the
bugreport.

For Bubble UI scenario for instance, we would see a log like this:

  ... imeDisplayId=0 ... targetDisplayId=1 ...

There should be no user-visible behavior change.

 [1]: Iedd71e4ddf4983f90b02dd72e471e7fa8e838fbf
      7613cac9696dfda6684ed65b0919dd9ee4e77c04
 [2]: Icb29c4ef0703553ead8654d3255a554de6d1a9b4
      788712f5099df58e007c1d8cc30a99788baef4e1

Bug: 35079353
Bug: 114350902
Test: Manually verified with Bubble test app and run 'adb shell
      dumpsys input_method' to see that a virtual display ID is logged
      when typing on Bubble UI.
Change-Id: I6f40d86a03a7965d72af7e9a1b6466350b7dfab7