Fix flaky testForAllWinddow
It was flaky because the exiting window may be removed before the
assertion. Prevent it from removal can fix it.

During the setup, while window get created, there are
WindowSurfacePlacer traversal being posted to mAnimationHandler. This
could happen while we run the test. During the traversal, all exiting
window will be removed if possible, and the test window will get
removed.

Wait until the traversal finished can fix the issue.

Bug: 77772044
Bug: 74078662
Bug: 124088319
Test: atest DisplayContentTests#testForAllWindows
Change-Id: I7f14605d3bd4a548755b7e88b0294168ab46ec7b