Broaden temporary workaround.
We have logic that automatically grants the legacy "Storage"
permission when the user grants either of the new "Music" or
"Photos & Videos" permissions, but that logic only lives in
PermissionController, and does not apply to pre-grants.

As a temporary workaround, explicitly grant the "Storage"
permission in pre-grants until we can refactor the
PermissionController logic.

Bug: 124766975, 124466734
Test: manual
Change-Id: I85cdf567af7683c0a3cfba755a9135ebad7a290f