Using linear interpolation in fragment shader to avoid over saturation.
The color of wallpaper is over saturated after remapping in the
fragment shader so we also apply linear interpolation between original
color and remapped color to avoid over saturation.

In addition, we also add some comments, which is offered by UX,
to fragment shader, thanks the great help of Vignesh!

Bug: 126500189
Bug: 123615467
Test: Manually set wallpaper, compare the saturation of the image between wallpaper and wallpaper picker.
Change-Id: Ibc063dfbc880c9b7dce8bec33fa2fd189ecf7d72