Disable any fullScreenIntent launches in SystemUI.
We are not using SystemUI's notifications, therefore it doesn't make sense to still let SystemUI do work in the background.
This also fixes b/126789972, because NotificationEntryManager is launching the full screen intent upon adding the notification (because none of our apps is immersive). In car, we'd never want notifications to trigger context switches.

Change-Id: Ic907605f015a3e1dc56bfbcbfea2268360322a44
Fix: 126789972
Test: Use the "CALL" button in Kitchensink's notification; the pending intent that jumps back to kitchen sink should not be launched.