Fix black screen when finish FallbackHome
The reason for the short black screen is that FallbackHome will be destroyed
directly when the it finish, and there will be no home activity until the
NexusLauncher resumed.
In order to fix this problem, we can check if the display also contains the
home stack, because in this case there must have a home activity on that display.

- Also add a unit test to verify the home stack shouldn't be destroy
immediately.

Fix: 119392020
Test: atest ActivityStackTests ActivityDisplayTests

Change-Id: Ic0035f784a6b29c212ebc33d8addfcb3f3bf38c6