Use FLAG_BUBBLE rather than setting the value ourselves
This flag will be the thing we use to know if something should be
in the bubble state, so lets get rid of this extra notion.

Previously we relied on BubbleController flagging the entry when
onPendingEntryAdded is called in the NotifEntryListener, now we can rely
on NoMan to tell us the right thing.

Also cleans up a little how createBubble works in NotifTestHelper

Bug: 129147774
Test: atest BubbleControllerTest
Change-Id: I828af8740ac9b62680b249d877da519a8a2271d0