Fix compressed stubs
Whenever the stub package is uninstalled from /data, we now
install the compressed version from the system image.

Disabling a package still does not remove any version on /data.
This means that a compressed package may be un-updated, yet
disabled and it will continue to consume space on /data. To
clear /data completely, disable the application and then uninstall.

Fixes: 123919310
Test: Manual. Put compressed APK on the /system image
Test: Boot the device, see two packages are installed [/system & /data]
Test: Uninstall the appliction on /data, see two packages are installed [/system & /data]. verified package on /data is new
Test: Disable the application, see two packages are installed [/system & data] and user-disabled. verified package on /data is same pre/post disable
Test: Uninstall the disabled application, see one package is installed [/system] and user-disabled
Test: Enable the application, see two packages are installed [/system & data] and enabled
Test: Corrupt compressed APK and reboot/enable/uninstall update, see one package is installed [/system] and disabled
Change-Id: I6fdd3ed485db0d3f8135266010d6797d9ac37ea0